java-lotto/yuus95 로또 게임 기능 구현

+++ src/main/java/lotto/LottoPrizeStrategy.java
@@ -3,5 +3,5 @@ package lotto;
 import java.util.List;
 
 public interface LottoPrizeStrategy {
-    int getMatch(Lotto lotto, List<Lotto> targetLottoList);
+    long getMatch(Lotto lotto, List<Lotto> targetLottoList);
 }
